backend/requirements.txt
aioapns==2.2
amqp==5.1.1
anyio==3.7.0
asammdf==7.3.12
asgiref==3.7.2
async-timeout==4.0.3
axios==0.1.0
channels==4.2.2
charset-normalizer==3.1.0
DateTime==5.2
dj-database-url==1.3.0
dj-static==0.0.6
Django==5.0.3
django-cors-headers==4.3.1
django-db-connection-pool==1.2.4
django-debug-toolbar==4.3.0
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-jwt==1.0.1
python-dotenv==1.0.0
mysqlclient==2.2.0
supervisor
paho-mqtt==2.1.0
daphne
channels_redis